refactor(core): extract createLabelInstanceManager
Pull the label/grabbed-box state machine out of init() into a focused
manager (PLAN §3.1.b). The manager owns the two timer maps
(labelFadeTimeouts, grabbedBoxTimeouts) and exposes a 6-method surface:

  - showTemporaryGrabbedBox
  - createLabelInstance
  - createPerElementLabelInstances
  - clearAllLabels
  - updateLabelAfterCopy
  - handleLabelInstanceHoverChange
  - dispose

The 'compute mouseX offsets' algorithm that was hand-rolled in both
createLabelInstance and createPerElementLabelInstances is now a single
computeMouseXOffsets() helper shared by both.

The dispose path now goes through labelManager.dispose() instead of
three separate timer-cleanup calls inside the api.dispose closure.

Also drops 3 now-unused imports (FADE_COMPLETE_BUFFER_MS, GrabbedBox
type, generateId).

index.tsx: 3579 -> 3423 lines (-156). Behaviour preserved:
copy-feedback.spec.ts + context-menu.spec.ts (57 tests) green.
